+++ api/src/main/java/org/allaymc/api/eventbus/event/world/WeatherChangeEvent.java
@@ -5,19 +5,17 @@ import org.allaymc.api.eventbus.event.CancellableEvent;
 import org.allaymc.api.world.World;
 import org.allaymc.api.world.data.Weather;
 
-import java.util.Set;
-
 /**
  * @author daoge_cmd
  */
 @Getter
 public class WeatherChangeEvent extends WorldEvent implements CancellableEvent {
-    protected Set<Weather> oldWeathers;
-    protected Set<Weather> newWeathers;
+    protected Weather oldWeather;
+    protected Weather newWeather;
 
-    public WeatherChangeEvent(World world, Set<Weather> oldWeathers, Set<Weather> newWeathers) {
+    public WeatherChangeEvent(World world, Weather oldWeather, Weather newWeather) {
         super(world);
-        this.oldWeathers = oldWeathers;
-        this.newWeathers = newWeathers;
+        this.oldWeather = oldWeather;
+        this.newWeather = newWeather;
     }
 }