+++ api/src/main/java/org/allaymc/api/world/chunk/FakeChunkLoader.java
@@ -12,7 +12,7 @@ import org.allaymc.api.entity.interfaces.EntityPlayer;
 import org.allaymc.api.math.location.Location3d;
 import org.allaymc.api.math.location.Location3dc;
 import org.allaymc.api.server.Server;
-import org.allaymc.api.world.World;
+import org.allaymc.api.world.data.Weather;
 import org.cloudburstmc.protocol.bedrock.packet.BedrockPacket;
 import org.joml.Vector3dc;
 import org.joml.Vector3ic;
@@ -138,7 +138,7 @@ public record FakeChunkLoader(Supplier<Location3dc> locationSupplier, int radius
     }
 
     @Override
-    public void viewTime(World world) {
+    public void viewTime(int timeOfDay) {
     }
 
     @Override
@@ -147,5 +147,13 @@ public record FakeChunkLoader(Supplier<Location3dc> locationSupplier, int radius
 
     @Override
     public void viewBlockUpdates(Chunk chunk, Collection<BlockUpdate> blockUpdates, Collection<BlockUpdate> extraBlockUpdates) {
+    }
+
+    @Override
+    public void viewSound(Vector3dc pos, String sound, double volume, double pitch) {
+    }
+
+    @Override
+    public void viewWeather(Weather weather) {
     }
 }