+++ api/src/main/java/org/allaymc/api/world/data/Weather.java
@@ -2,9 +2,6 @@ package org.allaymc.api.world.data;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import org.cloudburstmc.math.vector.Vector3f;
-import org.cloudburstmc.protocol.bedrock.data.LevelEvent;
-import org.cloudburstmc.protocol.bedrock.packet.LevelEventPacket;
 
 import java.util.concurrent.ThreadLocalRandom;
 
@@ -16,14 +13,12 @@ import java.util.concurrent.ThreadLocalRandom;
 @Getter
 @AllArgsConstructor
 public enum Weather {
-    CLEAR(20 * 60 * 10, 20 * 60 * 150, null, null),
-    RAIN(20 * 60 * 3, 20 * 60 * 13, LevelEvent.START_RAINING, LevelEvent.STOP_RAINING),
-    THUNDER(20 * 60 * 3, 20 * 60 * 13, LevelEvent.START_THUNDERSTORM, LevelEvent.STOP_THUNDERSTORM);
+    CLEAR(20 * 60 * 10, 20 * 60 * 150),
+    RAIN(20 * 60 * 3, 20 * 60 * 13),
+    THUNDER(20 * 60 * 3, 20 * 60 * 13);
 
     private final int minTime;
     private final int maxTime;
-    private final LevelEvent startLevelEvent;
-    private final LevelEvent stopLevelEvent;
 
     /**
      * Generate a random time length for this weather.
@@ -32,38 +27,5 @@ public enum Weather {
      */
     public int generateRandomTimeLength() {
         return ThreadLocalRandom.current().nextInt(minTime, maxTime + 1);
-    }
-
-    /**
-     * Create a start level event packet for this weather.
-     *
-     * @return the start level event packet for this weather, or {@code null} if the start level event is {@code null}
-     */
-    public LevelEventPacket createStartLevelEventPacket() {
-        if (startLevelEvent == null) {
-            return null;
-        }
-
-        var pk = new LevelEventPacket();
-        pk.setType(startLevelEvent);
-        pk.setPosition(Vector3f.ZERO);
-        pk.setData(65535);
-        return pk;
-    }
-
-    /**
-     * Create a stop level event packet for this weather.
-     *
-     * @return the stop level event packet for this weather, or {@code null} if the stop level event is {@code null}
-     */
-    public LevelEventPacket createStopLevelEventPacket() {
-        if (stopLevelEvent == null) {
-            return null;
-        }
-
-        var pk = new LevelEventPacket();
-        pk.setType(stopLevelEvent);
-        pk.setPosition(Vector3f.ZERO);
-        return pk;
     }
 }